Define Café/Coffee data to RAG Lite service
RagLite use case is a naive implementation to rag that inject all json into the
prompt and let the AI use it to interact.

The Data are organized as follow:

* data/cafes/catalogo_cafe_details.json
* data/cafes/catalog_cafe_details.toon
A Coffee catalog with detailed information about the coffees. Use it and your
token cost increase.

* data/cafes/catalog_cafe_small.json
* data/cafes/catalog_cafe_small.toon
A Coffee catalog short in information about the coffees. Use it to a more effective
token cost.

* data/cafes/categorias.json
A category Tree only about coffee.

* data/cafes/prompt.txt
The prompt used on this Rag Lite.